# Use memchr for str::find(char)