--- a/src/libstd/lib.rs
+++ b/src/libstd/lib.rs
@@ -302,6 +302,7 @@
 #![feature(sip_hash_13)]
 #![feature(slice_bytes)]
 #![feature(slice_concat_ext)]
+#![feature(slice_internals)]
 #![feature(slice_patterns)]
 #![feature(staged_api)]
 #![feature(stmt_expr_attributes)]