… src/libcore/tests/lib.rs
… src/libcore/tests/pattern.rs
… src/libstd/sys/unix/memchr.rs
… src/libstd/sys/wasm/memchr.rs
… src/libstd/sys/windows/memchr.rs
… src/libstd/sys_common/memchr.rs
… src/libstd/sys_common/mod.rs