This is a 10x improvement for searching for characters.


This also contains the patches from https://github.com/rust-lang/rust/pull/46713 . Feel free to land both separately or together.


cc @mystor @alexcrichton

r? @bluss

fixes #46693