@Manishearth 2017-12-14
I haven't really tested this much, there probably are failures. Will do a second pass at self-review once I know we pass all tests (from travis)