@Manishearth 2017-12-14
This does not bring improvements for multibyte chars or for `str::find(str)`. We _can_ bring improvement for these, but it's tricky.

For `str` when it starts with an ASCII char we can do similar stuff as here (and then use the original algorithm to finish the match.

When the thing we're searching for is _not_ ASCII we can still search for the first byte. However for most UTF8 text the first byte will generally be pretty uniform; i.e. if it's Arabic text will usually be 0xD8 or 0xD9, Korean will be 0xEA, 0xEB, 0xEC, or 0xED, Devanagari is usually 0xE0, etc. This means that memchr will have lots of false positives; we'll get lots of hits on the first byte and then have to check the second byte. This amount of stutter will probably make memchr's (minor) fixed overhead significant, and destroy any perf gains which we may get.

Searching for the second byte or even better, the _last_ byte, might work better. But I'm not sure if I want to write that code right now, and the tradeoffs are a bit trickier there :)